fabric/requirements.txt
Django==4.2.7
django-crispy-forms==2.3
crispy-bootstrap5==2023.10 